Carga_sqlserver/requirements.txt
pandas>=2.1
sqlalchemy>=2.0
pyodbc>=5.1
python-dotenv>=1.0
openpyxl>=3.1
tqdm>=4.66
